Refactor transformation classes to another file
