+++ dace/libraries/blas/nodes/dot.py
@@ -3,7 +3,7 @@ import dace.library
 import dace.properties
 import dace.sdfg.nodes
 from dace.symbolic import symstr
-from dace.transformation.pattern_matching import ExpandTransformation
+from dace.transformation.transformation import ExpandTransformation
 from dace.libraries.blas.nodes.matmul import _get_matmul_operands
 from .. import environments
 
@@ -19,13 +19,13 @@ class ExpandDotPure(ExpandTransformation):
 
         ((edge_x, outer_array_x, shape_x, _), (edge_y, outer_array_y, shape_y,
                                                _),
-         (_, outer_array_result, shape_result, _)) = _get_matmul_operands(
-             node,
-             parent_state,
-             parent_sdfg,
-             name_lhs="_x",
-             name_rhs="_y",
-             name_out="_result")
+         (_, outer_array_result, shape_result,
+          _)) = _get_matmul_operands(node,
+                                     parent_state,
+                                     parent_sdfg,
+                                     name_lhs="_x",
+                                     name_rhs="_y",
+                                     name_out="_result")
 
         dtype_x = outer_array_x.dtype.type
         dtype_y = outer_array_y.dtype.type
@@ -42,7 +42,9 @@ class ExpandDotPure(ExpandTransformation):
 
         _, array_x = sdfg.add_array("_x", shape_x, dtype_x, storage=storage)
         _, array_y = sdfg.add_array("_y", shape_y, dtype_y, storage=storage)
-        _, array_result = sdfg.add_array("_result", [1], dtype_result, storage=storage)
+        _, array_result = sdfg.add_array("_result", [1],
+                                         dtype_result,
+                                         storage=storage)
 
         mul_program = "__out = __x * __y"
 
@@ -54,13 +56,15 @@ class ExpandDotPure(ExpandTransformation):
 
         # Initialization map
         init_write = init_state.add_write("_result")
-        init_tasklet = init_state.add_tasklet(
-            "dot_init", {}, {"_out"}, "_out = 0", location=node.location)
-        init_state.add_memlet_path(
-            init_tasklet,
-            init_write,
-            src_conn="_out",
-            memlet=dace.Memlet.simple(init_write.data, "0", num_accesses=1))
+        init_tasklet = init_state.add_tasklet("dot_init", {}, {"_out"},
+                                              "_out = 0",
+                                              location=node.location)
+        init_state.add_memlet_path(init_tasklet,
+                                   init_write,
+                                   src_conn="_out",
+                                   memlet=dace.Memlet.simple(init_write.data,
+                                                             "0",
+                                                             num_accesses=1))
 
         # Multiplication map
         state.add_mapped_tasklet(
