+++ dace/libraries/blas/nodes/transpose.py
@@ -5,7 +5,7 @@ from dace.config import Config
 import dace.library
 import dace.properties
 import dace.sdfg.nodes
-from dace.transformation.pattern_matching import ExpandTransformation
+from dace.transformation.transformation import ExpandTransformation
 from .. import environments
