+++ dace/transformation/dataflow/double_buffering.py
@@ -8,13 +8,13 @@ from dace import data, dtypes, sdfg as sd, subsets, symbolic, registry
 from dace.memlet import Memlet
 from dace.sdfg import nodes
 from dace.sdfg import utils as sdutil
-from dace.transformation import pattern_matching
+from dace.transformation import transformation
 
 from dace.transformation.dataflow.map_for_loop import MapToForLoop
 
 
 @registry.autoregister_params(singlestate=True)
-class DoubleBuffering(pattern_matching.Transformation):
+class DoubleBuffering(transformation.Transformation):
     """ Implements the double buffering pattern, which pipelines reading
         and processing data by creating a second copy of the memory.
         In particular, the transformation takes a 1D map and all internal
@@ -81,8 +81,8 @@ class DoubleBuffering(pattern_matching.Transformation):
         map_rstart, map_rend, map_rstride = map_entry.map.range[0]
         map_rend = symbolic.pystr_to_symbolic('(%s) - (%s)' %
                                               (map_rend, map_rstride))
-        map_entry.map.range = subsets.Range([(map_rstart, map_rend,
-                                              map_rstride)])
+        map_entry.map.range = subsets.Range([(map_rstart, map_rend, map_rstride)
+                                             ])
 
         ##############################
         # Gather transients to modify
@@ -217,8 +217,8 @@ class DoubleBuffering(pattern_matching.Transformation):
             # Already in the right shape, modify new dimension
             subset = list(subset)[1:]
 
-        new_subset = subsets.Range([('__dace_db_param', '__dace_db_param',
-                                     1)] + list(subset))
+        new_subset = subsets.Range([('__dace_db_param', '__dace_db_param', 1)] +
+                                   list(subset))
         return new_subset
 
     @staticmethod
