+++ dace/transformation/dataflow/gpu_transform_local_storage.py
@@ -7,7 +7,7 @@ import copy
 from dace import data, dtypes, registry, sdfg as sd, subsets as sbs, symbolic
 from dace.sdfg import nodes
 from dace.sdfg import utils as sdutil
-from dace.transformation import pattern_matching
+from dace.transformation import transformation
 from dace.properties import Property, make_properties
 from dace.config import Config
 
@@ -35,7 +35,7 @@ def in_path(path, edge, nodetype, forward=True):
 
 @registry.autoregister_params(singlestate=True)
 @make_properties
-class GPUTransformLocalStorage(pattern_matching.Transformation):
+class GPUTransformLocalStorage(transformation.Transformation):
     """Implements the GPUTransformLocalStorage transformation.
 
         Similar to GPUTransformMap, but takes multiple maps leading from the
@@ -111,8 +111,8 @@ class GPUTransformLocalStorage(pattern_matching.Transformation):
             for node in subgraph.nodes():
                 if (isinstance(node, nodes.AccessNode) and
                         node.desc(sdfg).storage != dtypes.StorageType.Default
-                        and node.desc(sdfg).storage !=
-                        dtypes.StorageType.Register):
+                        and
+                        node.desc(sdfg).storage != dtypes.StorageType.Register):
                     return False
 
             # If one of the outputs is a stream, do not match
@@ -431,8 +431,7 @@ class GPUTransformLocalStorage(pattern_matching.Transformation):
                         memlet.data = node.data
 
                     if self.fullcopy:
-                        edge.data.subset = sbs.Range.from_array(
-                            node.desc(sdfg))
+                        edge.data.subset = sbs.Range.from_array(node.desc(sdfg))
                     edge.data.other_subset = newmemlet.subset
                     graph.add_edge(edge.src, edge.src_conn, node, None,
                                    edge.data)
@@ -486,8 +485,7 @@ class GPUTransformLocalStorage(pattern_matching.Transformation):
                             continue
                         path = graph.memlet_path(e)
                         if not isinstance(path[0].dst, nodes.CodeNode):
-                            if in_path(path, e, nodes.EntryNode,
-                                       forward=False):
+                            if in_path(path, e, nodes.EntryNode, forward=False):
                                 if isinstance(parent, nodes.CodeNode):
                                     # Output edge
                                     break
@@ -517,8 +515,7 @@ class GPUTransformLocalStorage(pattern_matching.Transformation):
 
                     edge.data.wcr = None
                     if self.fullcopy:
-                        edge.data.subset = sbs.Range.from_array(
-                            node.desc(sdfg))
+                        edge.data.subset = sbs.Range.from_array(node.desc(sdfg))
                     edge.data.other_subset = newmemlet.subset
                     graph.add_edge(node, None, edge.dst, edge.dst_conn,
                                    edge.data)
