+++ dace/transformation/dataflow/map_collapse.py
@@ -5,14 +5,14 @@ from dace import registry
 from dace.symbolic import symlist
 from dace.sdfg import nodes
 from dace.sdfg import utils as sdutil
-from dace.transformation import pattern_matching
+from dace.transformation import transformation
 from dace.properties import make_properties
 from typing import Tuple
 
 
 @registry.autoregister_params(singlestate=True)
 @make_properties
-class MapCollapse(pattern_matching.Transformation):
+class MapCollapse(transformation.Transformation):
     """ Implements the Map Collapse pattern.
 
         Map-collapse takes two nested maps with M and N dimensions respectively,
@@ -34,10 +34,8 @@ class MapCollapse(pattern_matching.Transformation):
     @staticmethod
     def can_be_applied(graph, candidate, expr_index, sdfg, strict=False):
         # Check the edges between the entries of the two maps.
-        outer_map_entry = graph.nodes()[candidate[
-            MapCollapse._outer_map_entry]]
-        inner_map_entry = graph.nodes()[candidate[
-            MapCollapse._inner_map_entry]]
+        outer_map_entry = graph.nodes()[candidate[MapCollapse._outer_map_entry]]
+        inner_map_entry = graph.nodes()[candidate[MapCollapse._inner_map_entry]]
 
         # Check that inner map range is independent of outer range
         map_deps = set()
@@ -88,10 +86,8 @@ class MapCollapse(pattern_matching.Transformation):
 
     @staticmethod
     def match_to_str(graph, candidate):
-        outer_map_entry = graph.nodes()[candidate[
-            MapCollapse._outer_map_entry]]
-        inner_map_entry = graph.nodes()[candidate[
-            MapCollapse._inner_map_entry]]
+        outer_map_entry = graph.nodes()[candidate[MapCollapse._outer_map_entry]]
+        inner_map_entry = graph.nodes()[candidate[MapCollapse._inner_map_entry]]
 
         return ' -> '.join(entry.map.label + ': ' + str(entry.map.params)
                            for entry in [outer_map_entry, inner_map_entry])
