+++ dace/transformation/dataflow/mapreduce.py
@@ -10,7 +10,7 @@ from dace.properties import Property, make_properties
 from dace.sdfg import SDFG
 from dace.sdfg import utils as sdutil
 from dace.symbolic import symstr
-from dace.transformation import pattern_matching as pm
+from dace.transformation import transformation as pm
 
 from dace.transformation.dataflow.map_collapse import MapCollapse
 from dace.transformation.dataflow.map_fusion import MapFusion
