+++ dace/transformation/dataflow/merge_arrays.py
@@ -1,5 +1,5 @@
 # Copyright 2019-2020 ETH Zurich and the DaCe authors. All rights reserved.
-from dace.transformation import pattern_matching
+from dace.transformation import transformation
 from dace import memlet, registry
 from dace.sdfg import nodes
 from dace.sdfg import SDFGState
@@ -7,7 +7,7 @@ from dace.sdfg.propagation import propagate_memlet
 
 
 @registry.autoregister_params(singlestate=True, strict=True)
-class InMergeArrays(pattern_matching.Transformation):
+class InMergeArrays(transformation.Transformation):
     """ Merge duplicate arrays connected to the same scope entry. """
 
     _array1 = nodes.AccessNode("_")
@@ -122,7 +122,7 @@ class InMergeArrays(pattern_matching.Transformation):
 
 
 @registry.autoregister_params(singlestate=True, strict=True)
-class OutMergeArrays(pattern_matching.Transformation):
+class OutMergeArrays(transformation.Transformation):
     """ Merge duplicate arrays connected to the same scope entry. """
 
     _array1 = nodes.AccessNode("_")
