+++ dace/transformation/interstate/fpga_transform_state.py
@@ -5,7 +5,7 @@ import dace
 from dace import data, memlet, dtypes, registry, sdfg as sd, subsets
 from dace.sdfg import nodes
 from dace.sdfg import utils as sdutil
-from dace.transformation import pattern_matching
+from dace.transformation import transformation
 
 
 def fpga_update(sdfg, state, depth):
@@ -30,7 +30,7 @@ def fpga_update(sdfg, state, depth):
 
 
 @registry.autoregister
-class FPGATransformState(pattern_matching.Transformation):
+class FPGATransformState(transformation.Transformation):
     """ Implements the FPGATransformState transformation. """
 
     _state = sd.SDFGState()
