+++ dace/transformation/subgraph/gpu_persistent_fusion.py
@@ -7,7 +7,7 @@ from dace.dtypes import StorageType, ScheduleType
 from dace.properties import Property, make_properties
 from dace.sdfg.utils import find_sink_nodes, concurrent_subgraphs
 from dace.sdfg.graph import SubgraphView
-from dace.transformation.pattern_matching import SubgraphTransformation
+from dace.transformation.transformation import SubgraphTransformation
 
 
 @registry.autoregister
