+++ dace/transformation/subgraph/helpers.py
@@ -3,7 +3,6 @@ from dace import dtypes, registry, symbolic, subsets
 from dace.sdfg import nodes, utils
 from dace.memlet import Memlet
 from dace.sdfg import replace, SDFG, SDFGState
-from dace.transformation import pattern_matching
 from dace.properties import make_properties, Property
 from dace.symbolic import symstr
 from dace.sdfg.propagation import propagate_memlets_sdfg
